message button

+++ resources/views/friend/index.blade.php
@@ -9,22 +9,34 @@
                 <h2 class="">{{$user->firstName}}</h2>
                 <h2 class="ml-3">{{$user->lastName}}</h2>
             </div>
-	        <div class="col-2">
-	            {{--friend詳細画面--}}
-                {!!link_to_route('users.show','check',['user'=>$user->id],['class'=>'btn btn-outline-secondary btn-block'])!!}
-            </div>
-            <div class="col-2">
-                {{--ユーザとのメッセージ画面--}}
-                {!!link_to_route('messages.show','message',['id'=>$user->id],['class'=>'btn btn-outline-primary btn-block'])!!} 
-               {{-- <a href="/messages/{{$user->id}}"><button type="button" class="btn btn-outline-primary btn-block">Message</button></a>--}}
-            </div>
-	        <div class="col-2">
-	            {{--friendをやめる--}}
-                {!!Form::open(['route'=>['friend.release','id'=>$user->id],'method'=>'delete'])!!}
-                {!!Form::submit('release',['class'=>'btn btn-outline-danger btn-block'])!!}
-                {!!Form::close()!!}
-                {{--{!!link_to_route('friend.release','release',['id'=>$user->id],['class'=>'btn btn-outline-danger btn-block'])!!}--}}
-            </div>
+            @if($id==Auth::id())
+	            <div class="col-2">
+                    {{--friend詳細画面--}}
+                    {!!link_to_route('users.show','check',['user'=>$user->id],['class'=>'btn btn-outline-secondary btn-block'])!!}
+                </div>
+                <div class="col-2">
+                    {{--ユーザとのメッセージ画面--}}
+                    {!!link_to_route('messages.show','message',['id'=>$user->id],['class'=>'btn btn-outline-primary btn-block'])!!} 
+                </div>
+	            <div class="col-2">
+	                {{--friendをやめる--}}
+                    {!!Form::open(['route'=>['friend.release','id'=>$user->id],'method'=>'delete'])!!}
+                    {{--{!!link_to_route('friend.release','release',['id'=>$user->id],['class'=>'btn btn-outline-danger btn-block'])!!}--}}
+                    {!!Form::submit('release',['class'=>'btn btn-outline-danger btn-block'])!!}
+                    {!!Form::close()!!}
+                </div>
+            @else
+	            <div class="col-2">
+                    {{--friend詳細画面--}}
+                    {!!link_to_route('users.show','check',['user'=>$user->id],['class'=>'btn btn-outline-secondary btn-block'])!!}
+                </div>
+                @if($user->id!=Auth::id())
+                    <div class="col-2">
+                        {{--ユーザとのメッセージ画面--}}
+                        {!!link_to_route('messages.show','message',['id'=>$user->id],['class'=>'btn btn-outline-primary btn-block'])!!} 
+                    </div>
+                @endif
+            @endif
         </div>
     @endforeach
 @else
